+++ src/main/java/ru/tn/courses/kskubnikov/task1/prob1/Main.java
@@ -1,4 +1,4 @@
-package task3;
+package task1.prob1;
 
 import java.util.Arrays;
 import java.util.Random;
@@ -8,25 +8,20 @@ import java.util.Random;
  */
 public class Main {
 
-    private static int arr[];
-    private static int SIZE = 15;
+    private static int[ ] arr;
+    private static final int SIZE = 10;
+    private static boolean flag = true;
 
     public static void main(String[] args) {
         arrInit();
-        System.out.println(Arrays.toString(arr));
-
-        for(int i = 1; i < arr.length; i++){
-            int curr = arr[i];
-            int j = i-1;
-            while(j >= 0 && arr[j] > curr){
-                arr[j+1] = arr[j];
-                j--;
+        for (int i = 0; i < arr.length - 1; i++) {
+            if(arr[i] > arr[i+1]){
+                flag = false;
+                break;
             }
-            arr[j+1] = curr;
         }
-
         System.out.println(Arrays.toString(arr));
-
+        System.out.println("Incr?: " + flag );
     }
 
     private static void arrInit(){
